chore: fix a build failure on kpathsea with emscripten
The build will fail if compiled with emscripten by:

```
emconfigure ./configure
emmake make
```

XeTeX's kpathsea library seems to be outdated which lacks a fix from
[here](https://github.com/TeX-Live/texlive-source/commit/1aae4f2a50fc8c84e187914056205010ac1ec70f)

However, the build will succeed if build with normal gcc:

```
./configure
make
```

The root cause for this here is, when you execute `./configure`, the
value of `GETCWD_FORKS` is defined as in `c-auto.h`:

```c
/* Define to 1 if getcwd is implemented using fork or vfork. Let me know if
   you have to add this by hand because configure failed to detect it. */
/* #undef GETCWD_FORKS */
```

However, when you execute `emconfigure ./configure`, the value of
`GETCWD_FORKS` is defined as:

```c
/* Define to 1 if getcwd is implemented using fork or vfork. Let me know if
   you have to add this by hand because configure failed to detect it. */
```

Thus the code in `xgetcwd.c` will be compiled and expanded with
different part:

```c
```

That is why gcc succeed while emcc reports error here.

However, the reason of why `emconfigure ./configure` and `./configure`
produce different `GETCWD_FORKS`s is still unknown to me, but I don't
want to spend more time on this though.